# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateMemoryRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "editor": {
            "title": "Editor",
            "type": "string"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "kind": {
            "enum": [
              "fact",
              "preference",
              "procedure",
              "project_note",
              "persona",
              "pinned"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "string"
          },
          "project_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Key"
          },
          "thread_origin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thread Origin"
          }
        },
        "required": [
          "principal_id",
          "label",
          "body",
          "kind",
          "editor"
        ],
        "title": "CreateMemoryRequest",
        "type": "object"
      },
...
      "PatchMemoryRequest": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "editor": {
            "title": "Editor",
            "type": "string"
          },
          "expected_revision": {
            "title": "Expected Revision",
            "type": "integer"
          },
          "keywords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Keywords"
          },
          "kind": {
            "anyOf": [
              {
                "enum": [
                  "fact",
                  "preference",
                  "procedure",
                  "project_note",
                  "persona",
                  "pinned"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "pin": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pin"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "enum": [
                  "active",
                  "quarantined",
                  "tombstoned"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "required": [
          "expected_revision",
          "editor",
          "reason"
        ],
        "title": "PatchMemoryRequest",
        "type": "object"
      },
...
      "SearchRequest": {
        "additionalProperties": false,
        "properties": {
          "k": {
            "default": 10,
            "title": "K",
            "type": "integer"
          },
          "principal_id": {
            "title": "Principal Id",
            "type": "string"
          },
          "project_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Key"
          },
          "query": {
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "principal_id",
          "query"
        ],
        "title": "SearchRequest",
        "type": "object"
      }
    },
    "securitySchemes": {
      "StaticBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "N8 Spine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/memories": {
      "get": {
        "operationId": "list_memories_v1_memories_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Key"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "active",
                    "quarantined",
                    "tombstoned"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          }
        ],
        "responses": {
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Bearer token missing or invalid"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Request does not match the endpoint contract"
          },
          "501": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Agent Zero contract stub"
          }
        },
        "security": [
          {
            "StaticBearer": []
          }
        ],
        "summary": "List Memories",
        "tags": [
          "memory"
        ]
      },
      "post": {
        "operationId": "create_memory_v1_memories_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Bearer token missing or invalid"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Request does not match the endpoint contract"
          },
          "501": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Agent Zero contract stub"
          }
        },
        "security": [
          {
            "StaticBearer": []
          }
        ],
        "summary": "Create Memory",
        "tags": [
          "memory"
        ]
      }
    },
    "/v1/memories/{id}": {
      "patch": {
        "operationId": "patch_memory_v1_memories__id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Bearer token missing or invalid"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Request does not match the endpoint contract"
          },
          "501": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Agent Zero contract stub"
          }
        },
        "security": [
          {
            "StaticBearer": []
          }
        ],
        "summary": "Patch Memory",
        "tags": [
          "memory"
        ]
      }
    },
    "/v1/search": {
      "post": {
        "operationId": "search_v1_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Bearer token missing or invalid"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Request does not match the endpoint contract"
          },
          "501": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "additionalProperties": true,
                  "description": "RFC 7807 problem details with an optional endpoint extension.",
                  "properties": {
                    "detail": {
                      "title": "Detail",
                      "type": "string"
                    },
                    "endpoint": {
                      "anyOf": [
                        {
                          "type": "string"
                        },
                        {
                          "type": "null"
                        }
                      ],
                      "title": "Endpoint"
                    },
                    "instance": {
                      "title": "Instance",
                      "type": "string"
                    },
                    "status": {
                      "title": "Status",
                      "type": "integer"
                    },
                    "title": {
                      "title": "Title",
                      "type": "string"
                    },
                    "type": {
                      "default": "about:blank",
                      "title": "Type",
                      "type": "string"
                    }
                  },
                  "required": [
                    "title",
                    "status",
                    "detail",
                    "instance"
                  ],
                  "title": "ProblemDetail",
                  "type": "object"
                }
              }
            },
            "description": "Agent Zero contract stub"
          }
        },
        "security": [
          {
            "StaticBearer": []
          }
        ],
        "summary": "Search",
        "tags": [
          "memory"
        ]
      }
    }
  }
}